+++ docs/src/release-notes.rst
@@ -7,6 +7,7 @@ Unreleased
 - Add multi-file, recursive and importable module analysis (:issue:`5`)
 - Split CLI file name specification to separate arguments (:issue:`5`)
 - Add option to specify graph root (:issue:`14`)
+- Expand assignment parsing (:issue:`18`)
 
 0.3.0 (2021-03-05)
 ------------------